move strings cut into nut and remove nut_assembly
